testcase controller test cases for success and object not found are all created and passing

+++ src/main/java/com/aphids/testplanner/service/TestcaseServiceImpl.java
@@ -47,10 +47,12 @@ public class TestcaseServiceImpl implements TestcaseService {
 
     @Override
     public void deleteTestcase(Long testcaseId) {
-
-        testcaseRepository.deleteById(testcaseId);
-
+        Optional<Testcase> testcaseOptional = testcaseRepository.findById(testcaseId);
+        if (testcaseOptional.isPresent()) {
+            testcaseRepository.deleteById(testcaseId);
+        } else {
+            throw new ObjectNotFoundException("Testcase not found: " + testcaseId);
+        }
     }
-
 
 }
